feat(product): controller for deleteProduct

+++ rest/src/controllers/productController.js
@@ -3,7 +3,7 @@ import { ProductModel } from "../models/Product.js";
 import { ManufacturerModel } from "../models/Manufacturer.js";
 import { ContactModel } from "../models/Contact.js";
 
-// Create a new Product
+// Create a new product
 export const createProduct = async (req, res) => {
   const { name, sku, description, price, category, manufacturer,  amountInStock } = req.body;
 
@@ -11,7 +11,7 @@ export const createProduct = async (req, res) => {
     return res.status(400).json({ error: `All fields are required ${Object.values( req.body ).join(", ")}` });
   }
 
-  if (!mongoose.Types.ObjectId.isValid(manufacturer)) {
+  if (!mongoose.isValidObjectId(manufacturer)) {
     return res.status(400).json({ error: "Manufacturer ID must be a valid ObjectID" });
   }
 
@@ -42,7 +42,7 @@ export const createProduct = async (req, res) => {
   }
 };
 
-// Get all users
+// Get all products
 export const getProducts = async (req, res) => {
   try {
     const products = await ProductModel.find()
@@ -54,10 +54,6 @@ export const getProducts = async (req, res) => {
           select: "name email"
         }
       }).lean();
-
-    if (!products.length) {
-      return res.status(404).json({ error: "No products found" });
-    }
 
     return res.status(200).json(products);
   } catch (err) {
@@ -102,7 +98,7 @@ export const updateProduct = async (req, res) => {
     return res.status(400).json({ error: "Product ID must be valid ObjectID" });
   }
 
-  if (!mongoose.Types.ObjectId.isValid(manufacturer)) {
+  if (!mongoose.isValidObjectId(manufacturer)) {
     return res.status(400).json({ error: "Manufacturer ID must be a valid ObjectID" });
   }
 
@@ -117,27 +113,39 @@ export const updateProduct = async (req, res) => {
   }
 
   try {
-    const product = await ProductModel.findByIdAndUpdate(
+    const updatedProduct = await ProductModel.findByIdAndUpdate(
       req.params.id,
       req.body,
-      { new: true, runValidators: true}
+      { new: true, runValidators: true }
     );
-    res.status(200).json(product);
+    return res.status(200).json({
+      message: "Product updated successfully",
+      product: updatedProduct
+    });
   } catch (error) {
     console.error(`[products/${req.params.id}]`, error);
-    res.status(500).json({ error: "Failed to update product" });
+    return res.status(500).json({ error: "Failed to update product" });
   }
 };
 
-// Delete a product
+// Delete a product by ID
 export const deleteProduct = async (req, res) => {
   if (!mongoose.isValidObjectId(req.params.id)) {
-    return res.status(400).json({ error: "ID must be valid ObjectID" });
+    return res.status(400).json({ error: "Product ID must be valid ObjectID" });
   }
+
+  const existingProduct = await ProductModel.findById(req.params.id);
+  if (!existingProduct) {
+    return res.status(404).json({ error: "Product not found" });
+  }
+
   try {
-    const products = await ProductModel.findByIdAndDelete(req.params.id);
-    res.json(products);
+    await ProductModel.findByIdAndDelete(req.params.id);
+    return res.status(200).json({
+      message: "Product deleted successfully",
+      product: existingProduct
+    });
   } catch (err) {
-    res.status(500).json({ error: err.message });
+    return res.status(500).json({ error: "Failed to delete product" } );
   }
 };
